requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
torch
torchvision
timm
mobile-sam